+++ tests/UninstallTest.php
@@ -3,6 +3,7 @@
 use Brain\Monkey;
 use Brain\Monkey\Functions;
 use FileToWeb\Integration\Document_State;
+use FileToWeb\Integration\PDF_To_Page;
 use FileToWeb\Integration\Settings;
 use PHPUnit\Framework\TestCase;
 
@@ -66,8 +67,10 @@ class UninstallTest extends TestCase {
 		$this->assertContains( Settings::OPTION_SETTINGS, $deleted_options );
 			$this->assertContains( Settings::LEGACY_OPTION_API_KEY, $deleted_options );
 			$this->assertContains( 'filetoweb_integration_bulk_queue', $deleted_options );
+			$this->assertContains( PDF_To_Page::OPTION_JOBS, $deleted_options );
 			$this->assertContains( Document_State::META_HTML_URL, $GLOBALS['wpdb']->deleted_meta_keys );
 			$this->assertContains( Document_State::META_LOCAL_HTML_PATH, $GLOBALS['wpdb']->deleted_meta_keys );
+			$this->assertContains( Document_State::META_PDF_TO_PAGE, $GLOBALS['wpdb']->deleted_meta_keys );
 			$this->assertContains( 'filetoweb_integration_poll_pending', $cleared_hooks );
 			$this->assertContains( 'filetoweb_integration_process_bulk_queue', $cleared_hooks );
 		}